rudimentary coin bullish/bearish states working

+++ exchange/exchange.go
@@ -2,10 +2,10 @@ package exchange
 
 type Exchange struct {
     name string
-    volume float32
+    volume float64
 }
 
-func New(name string, volume float32) Exchange {
+func New(name string, volume float64) Exchange {
     e := Exchange {name, volume}
     return e
 }
@@ -14,6 +14,6 @@ func (e Exchange) Name() string {
     return e.name
 }
 
-func (e Exchange) Volume() float32 {
+func (e Exchange) Volume() float64 {
     return e.volume
 }
